requirements.txt
streamlit
sentence-transformers
faiss-cpu
rank_bm25
pypdf
requests